server/requirements.txt
fastapi==0.115.0
uvicorn[standard]>=0.31.1
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
aiomqtt==2.3.0
pydantic>=2.11.0
pydantic-settings==2.5.2
scikit-learn==1.5.2
xgboost>=2.0.0
numpy==1.26.4
python-dotenv==1.0.1
rich>=13.0
requests>=2.31.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiosqlite>=0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
apscheduler>=3.10.0
pytest-cov>=5.0.0
mcp>=1.26.0
sse-starlette>=2.0.0
openai>=1.50.0
github-copilot-sdk>=0.2.0
